requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.7.1

numpy==2.4.1
matplotlib==3.10.8
opencv-python==4.13.0.90
imageio==2.37.2
pyyaml==6.0.3
einops==0.8.1